修改注解方式底层逻辑，测试添、删、查、改。 Signed-off-by: yanziqi <yanziq@caifu500.com>

+++ src/main/java/com/wyrz/core/dao/demain/Identifiable.java
@@ -1,7 +1,6 @@
 package com.wyrz.core.dao.demain;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 /**
  * 主键标识
@@ -18,7 +17,7 @@ public interface Identifiable extends Serializable {
 	 * @return
 	 * @date 2014年12月9日下午2:16:06
 	 */
-	public BigInteger getId();
+	public Integer getId();
 
 	/**
 	 * 设置主键标识
@@ -27,6 +26,6 @@ public interface Identifiable extends Serializable {
 	 * @param id
 	 * @date 2014年12月9日下午2:16:17
 	 */
-	public void setId(BigInteger id);
+	public void setId(Integer id);
 
 }
